chore: refactor image hooks + reconnect the new hook with components/pages
